\name{macro.pfsi-package}
\alias{macro.pfsi-package}
\alias{macro.pfsi}
\docType{package}
\title{
\packageTitle{macro.pfsi}
}
\description{
\packageDescription{macro.pfsi}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{macro.pfsi}
\packageIndices{macro.pfsi}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{macro.pfsi}

Maintainer: \packageMaintainer{macro.pfsi}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
